♻️ Move UserProvider and user hooks into features/user/client.tsx
