+++ apps/web/src/app/[locale]/(space)/layout.tsx
@@ -2,8 +2,8 @@ import { SessionRefresher } from "@/components/session-refresher";
 import { PayWall } from "@/features/billing/components/pay-wall";
 import { SpaceProvider } from "@/features/space/client";
 import { getActiveSpace } from "@/features/space/data";
+import { UserProvider } from "@/features/user/client";
 import { TimeZoneMismatchDialog } from "@/features/user/components/timezone-mismatch-dialog";
-import { UserProvider } from "@/features/user/components/user-provider";
 import { getLocale } from "@/i18n/server/get-locale";
 import { getSession } from "@/lib/auth";
 import { DateTimeProvider } from "@/lib/datetime/client";
