+++ apps/web/src/app/[locale]/(space)/settings/components/sidebar.tsx
@@ -26,7 +26,7 @@ import {
 import { usePathname } from "next/navigation";
 import { HoverPrefetchLink } from "@/components/hover-prefetch-link";
 import { useSpace } from "@/features/space/client";
-import { useAuthedUser } from "@/features/user/components/user-provider";
+import { useAuthedUser } from "@/features/user/client";
 import { Trans, useTranslation } from "@/i18n/client";
 import { useFeatureFlag } from "@/lib/feature-flags/client";
