+++ apps/web/src/app/[locale]/control-panel/layout.tsx
@@ -4,7 +4,7 @@ import { GaugeIcon } from "lucide-react";
 import type { Metadata } from "next";
 import { LicenseLimitWarning } from "@/features/licensing/components/license-limit-warning";
 import { CommandMenu } from "@/features/navigation/components/command-menu";
-import { UserProvider } from "@/features/user/components/user-provider";
+import { UserProvider } from "@/features/user/client";
 import { Trans } from "@/i18n/client";
 import { getTranslation } from "@/i18n/server";
 import { getLocale } from "@/i18n/server/get-locale";
