+++ apps/web/src/features/user/components/nav-user.tsx
@@ -30,7 +30,7 @@ import { useRouter } from "next/navigation";
 import React from "react";
 import { OptimizedAvatarImage } from "@/components/optimized-avatar-image";
 import { RouterLoadingIndicator } from "@/components/router-loading-indicator";
-import { useUser } from "@/features/user/components/user-provider";
+import { useUser } from "@/features/user/client";
 import { Trans } from "@/i18n/client";
 import { signOut } from "@/lib/auth-client";
 import { useTheme } from "@/lib/theme";
